requirements.txt
# Мультимодальный Braindler & Mozgach - Зависимости
# © 2025 NativeMind - NativeMindNONC License

# ============================================================
# Основные ML библиотеки
# ============================================================
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0

# ============================================================
# Vision (обработка изображений)
# ============================================================
Pillow>=10.0.0
opencv-python>=4.8.0
timm>=0.9.0  # для дополнительных vision моделей

# ============================================================
# OCR (для юридического режима)
# ============================================================
pytesseract>=0.3.10
easyocr>=1.7.0
paddleocr>=2.7.0  # альтернативный OCR
pdf2image>=1.16.3
PyMuPDF>=1.23.0  # fitz - для работы с PDF

# ============================================================
# Аудио (для будущего развития)
# ============================================================
librosa>=0.10.0
soundfile>=0.12.0
torchaudio>=2.0.0

# ============================================================
# Видео (для будущего развития)
# ============================================================
moviepy>=1.0.3
av>=10.0.0

# ============================================================
# Утилиты
# ============================================================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# ============================================================
# Обнаружение копипаста
# ============================================================
difflib  # стандартная библиотека
python-Levenshtein>=0.21.0  # быстрое сравнение строк
fuzzywuzzy>=0.18.0  # нечеткое сопоставление
jellyfish>=1.0.0  # фонетическое сравнение

# ============================================================
# HuggingFace интеграция
# ============================================================
huggingface-hub>=0.19.0
safetensors>=0.4.0

# ============================================================
# Мобильная оптимизация (опционально)
# ============================================================
# coremltools>=7.0  # для iOS
# tensorflow>=2.14.0  # для TFLite
# onnx>=1.15.0
# onnxruntime>=1.16.0

# ============================================================
# Веб и API (опционально)
# ============================================================
# fastapi>=0.104.0
# uvicorn>=0.24.0
# gradio>=4.0.0

# ============================================================
# Разработка и тестирование
# ============================================================
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0

